Solve Y2023/D05P1 and change solutions to u64

+++ src/solvers.rs
@@ -18,7 +18,7 @@ pub mod y2023;
 /// Solution returned by the solver
 #[derive(Debug, Display, PartialEq, Eq, Clone)]
 pub enum Solution {
-    Value(u32),
+    Value(u64),
     Err(String),
     NotExecuted,
     NotImplemented,
@@ -26,7 +26,7 @@ pub enum Solution {
 
 #[derive(Debug, Display, PartialEq, Eq, Clone)]
 pub enum SolutionExecution {
-    Value(u32),
+    Value(u64),
     NotImplemented,
 }
